moved fliters items fetching to the page level
